+++ website/src/routes/_benchmarks/_runtime/codec/-index.e2e.model.ts
@@ -1,4 +1,4 @@
-import { RuntimePage } from "#e2e/fixtures/pages/_runtime";
+import { RuntimePage } from "#e2e/fixtures/runtime.ts";
 
 export class CodecPage extends RuntimePage {
   url = "/codec";
