+++ website/src/routes/_benchmarks/_runtime/initialization/-index.e2e.model.ts
@@ -1,4 +1,4 @@
-import { RuntimePage } from "#e2e/fixtures/pages/_runtime";
+import { RuntimePage } from "#e2e/fixtures/runtime.ts";
 
 export class InitializationPage extends RuntimePage {
   url = "/initialization";
