+++ website/src/routes/_benchmarks/_runtime/standard/-index.e2e.model.ts
@@ -1,4 +1,4 @@
-import { RuntimePage, withDataToggle, withErrorTypeFilter } from "#e2e/fixtures/pages/_runtime";
+import { RuntimePage, withDataToggle, withErrorTypeFilter } from "#e2e/fixtures/runtime.ts";
 
 export class StandardSchemaPage extends withErrorTypeFilter(withDataToggle(RuntimePage)) {
   url = "/standard";
