+++ website/src/routes/_benchmarks/_runtime/validation/-index.e2e.model.ts
@@ -1,4 +1,4 @@
-import { RuntimePage, withDataToggle } from "#e2e/fixtures/pages/_runtime";
+import { RuntimePage, withDataToggle } from "#e2e/fixtures/runtime.ts";
 
 export class ValidationPage extends withDataToggle(RuntimePage) {
   url = "/validation";
